requirements.txt
# Core dependencies
python-telegram-bot>=20.0
google-generativeai>=0.3.0
gspread>=5.0
python-dotenv>=1.0.0
google-auth>=2.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
